# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/Seats": {
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Seat"
                            }
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "description": "Seat JSON",
                        "name": "Seat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    }
                }
            }
        },
        "/Seats/{id}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Seat"
                        }
                    }
                }
            }
        },
        "/movies": {
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Movie"
                            }
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "description": "Movie JSON",
                        "name": "Movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            },
            "patch": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                }
            }
        },
        "/shows": {
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Show"
                            }
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "description": "Show JSON",
                        "name": "Show",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    }
                }
            }
        },
        "/shows/{id}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    }
                }
            },
            "patch": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Show JSON",
                        "name": "Show",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    }
                }
            }
        },
        "/theaters": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Theater"
                            }
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "description": "Theater JSON",
                        "name": "Theater",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TheaterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Theater"
                        }
                    }
                }
            }
        },
        "/theaters/{id}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Theater"
                        }
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Theater"
                        }
                    }
                }
            },
            "patch": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Theater JSON",
                        "name": "Theater",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Theater"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Theater"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Movie": {
            "type": "object",
            "required": [
                "language",
                "length",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "length": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Seat": {
            "type": "object",
            "required": [
                "number",
                "theater_id",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "theater_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.SeatRequest": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Show": {
            "type": "object",
            "required": [
                "dateTime",
                "movieID",
                "theaterID"
            ],
            "properties": {
                "dateTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "movieID": {
                    "type": "string"
                },
                "theaterID": {
                    "type": "string"
                }
            }
        },
        "models.Theater": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Seat"
                    }
                }
            }
        },
        "models.TheaterRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "seat": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SeatRequest"
                    }
                }
            }
        }
    }
}